Adds parser for encryption dict
Signed-off-by: sameedali <sameed.ali.94@gmail.com>

This commit contains the code which reads the "/Encrypt" key in the
trailer of the pdf file, finds the object the "/Encrypt" key points to
(i.e. the encryption dictionary) and parses it.

This parsing done before XrefStm is parsed. This allows the ecnryption
code to be inserted later which will ensure that the streams are not
encrypted when the XrefStm parser is run.  This commit does *not*
implement the encryption algorithms. It only parses the encryption
dictionary and reads the values there in which will be used by a later
commit to implement the encryption algorithms.

On branch encryption-support
Changes to be committed:
      modified:   src/bin/pdf_printer.rs